backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic==0.39.0
scikit-learn==1.5.2
numpy==2.1.3
pydantic==2.10.0
python-dotenv==1.0.1
aiofiles==24.1.0
